Support Slack's markdown dialect
By allowing mrkdwn to be used in the message field. I used the
"markdown" syntax of the code editor field, though the syntax difference
is maybe too large for this resulting in the code editor displaying the
wrong style in quite a few cases. Of course I had no luck finding mrkdwn
code support for CodeMirror, it seems like it's not widely used outside
of Slack.